new module pushbtn to accept "gestures" from a push button switch and control an LED (the two are meant to work in a coordinated way, but can be used independently).
The test script works with an switch wired to GPIO18, and an LED wired
to GPIO17.  It has these semantics:
- click: toggle LED state
- doubleClick: show a slow on-off-on-off blink sequence
- hold: turn on the LED

Also, the test script subscribes to vahData messages from the master
nodejs process, so that the LED visualizes pulse/tag detections with
rapid blinks.

All of this is intended to allow "closed-box" operation of the RPi3
(and RPIzw, when it becomes available) sensorgnome.

The planned semantics:
- a single button push turns on (or off) "pulse/tag monitoring mode" on the LED
  (with a 15 minute timeout)
- a double button push turns on (or off) the WiFi hotspot in the SG
  (and 1/s heartbeat to show it is on) (with a 1 hour timeout).
- a 3s button hold turns on the LED and does a clean shutdown, turning
  off the LED when it is safe to remove the SD card