use CLOCK_MONOTONIC to measure timeout for detecting the GPS cape.
Detecting using CLOCK_REALTIME was sometimes failing due to a race
condition between GPS detection and setting of the system clock from
the GPS by gpsd and chrony, apparently.

Should this be fixed by some systemd wrangling to ensure the cape is
detected before use?  gpsd won't do anything wrong if the cape is
not present, as far as I can tell.

Closes #49