follow existing symlinks when unpacking un update
e.g. if an update inludes this file:

   /boot/uboot/something

and so its parent folders, then that file should end up in the existing folder reached by the
path /boot/uboot, which is really /dev/sdcard/uboot via
the symlink /boot -> /dev/sdcard
But without the 'h' option to tar, the symlink "/boot" is overwritten
by a new folder of that name.